Remove orgaRestAPI.java
